+++ src/main/java/edu/ucar/unidata/rosetta/service/validators/user/CommonUserValidator.java
@@ -1,4 +1,4 @@
-package edu.ucar.unidata.rosetta.service.user.validators;
+package edu.ucar.unidata.rosetta.service.validators.user;
 
 import edu.ucar.unidata.rosetta.service.user.UserManager;
 import edu.ucar.unidata.rosetta.service.validators.CommonValidator;
