+++ src/main/java/edu/ucar/unidata/rosetta/service/validators/user/EditUserValidator.java
@@ -1,4 +1,4 @@
-package edu.ucar.unidata.rosetta.service.user.validators;
+package edu.ucar.unidata.rosetta.service.validators.user;
 
 import edu.ucar.unidata.rosetta.domain.user.User;
 import edu.ucar.unidata.rosetta.service.user.UserManager;
